+++ src/selectors/station.selectors.js
@@ -1,0 +1,6 @@
+
+export function selectStationList (state) {
+  return Object.keys(state.station.byUuid).map((key, index) => {
+    return state.station.byUuid[key]
+  })
+}