+++ src/selectors/trails.selectors.js
@@ -1,4 +1,5 @@
 import { createSelector } from 'reselect'
+import { selectStationList } from './station.selectors.js'
 
 function selectTrailsList (state) {
   return Object.keys(state.trails.byUuid).map((key, index) => {
@@ -22,4 +23,9 @@ function selectTrailsByUuid (state) {
 export const getCurrentTrail = createSelector(
   [selectTrailsByUuid, selectCurrentTrailId],
   (trails, uuid) => trails[uuid]
+)
+
+export const getCurrentTrailsStations = createSelector(
+  [selectTrailsByUuid, selectCurrentTrailId, selectStationList],
+  (trails, current, stations) => stations.filter(s => trails[current].stations.includes(s.uuid))
 )